+++ src/user/UserDTO.java
@@ -3,6 +3,7 @@ package src.user;
 public class UserDTO {
 
     // 필드 선언
+    private int id;
     private String userId;
     private String userPw;
     private String userName;
@@ -21,6 +22,17 @@ public class UserDTO {
         this.userPhone = userPhone;
         this.userBirth = userBirth;
         this.userEmail = userEmail;
+    }
+
+    public UserDTO() {
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
     }
 
     public String getUserId() {
