+++ src/user/UserMain.java
@@ -4,10 +4,11 @@ public class UserMain {
 
     // 로그인 상태 유지
     public static String loginId;
+    public static int id;
 
     public static void main(String[] args) {
-        Controller controller = new Controller();
+        MainController controller = new MainController();
 
-        controller.firstMain();
+        controller.startMain();
     }
 }
